+++ CHANGELOG.md
@@ -76,6 +76,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#106](https://github.com/ARGOeu/argo-mon-status-api/pull/106) ARGO-5419 ARGO-5432 ARGO-5433 CRUD topology
 - [#108](https://github.com/ARGOeu/argo-mon-status-api/pull/108) ARGO-5443 Support node info in tenants and set reports as default
 - [#112](https://github.com/ARGOeu/argo-mon-status-api/pull/112) ARGO-5461: Sync Web API Tenants to Local DB (Dev Environment Only)
+- [#115](https://github.com/ARGOeu/argo-mon-status-api/pull/115) ARGO-5477: Support force parameter when creating new topology items
 
 ### Fix